add multi targeting frameworks in the same project for net 4.5 through net 7.0 and net standard 2.0
