+++ src/UglyToad.PdfPig/Fonts/SystemFonts/ISystemFontLister.cs
@@ -1,0 +1,9 @@
+﻿namespace UglyToad.PdfPig.Fonts.SystemFonts
+{
+    using System.Collections.Generic;
+
+    internal interface ISystemFontLister
+    {
+        IEnumerable<SystemFontRecord> GetAllFonts();
+    }
+}
