+++ src/UglyToad.PdfPig/Fonts/SystemFonts/SystemFontType.cs
@@ -1,0 +1,12 @@
+﻿namespace UglyToad.PdfPig.Fonts.SystemFonts
+{
+    internal enum SystemFontType
+    {
+        Unknown = 0,
+        TrueType = 1,
+        OpenType = 2,
+        Type1 = 3,
+        TrueTypeCollection = 4,
+        OpenTypeCollection = 5
+    }
+}
